CHG: Upgrade to qunit 2.0 CLI
Follows-up d23d4becc, which updated the package name for the
legacy node-qunit CLI wrapper. One of the reasons for that rename
was that the 'qunit' package name would now be used by QUnit itself,
which has a modern CLI built-in, largley inspired by node-qunit.

I've also upgraded the test suites to the 2.x API while at it,
which in this case really just means using QUnit.test() instead
of test(), but for simplicity I've kept the code as-is and just
defined that locally.

<https://qunitjs.com/upgrade-guide-2.x/>